test: split PDF prose continuity cases
